Refactor config (#288)
The following commits were squashed :

* refactor config.py to separate target specific code

- split config.py to DeployConfig base class and Ec2DeployConfig,
  ExistingDeployConfig, and AzureDeployConfig subclasses
- add @abstractmethod annotations for common interfaces to DeployConfig
  base class
- move target specific code to respective subclasses

* move config tests to target specific subdirectory

* add config decorators and move ec2 specific config

- add python decorators to annotate config items to allow generic
  handling of configs
- moved some ec2 specific configs to Ec2DeployConfig

* add config value validation decorator

* rename host_vars and play_vars to ansible_*

* use config/base.ansible_*_vars to render vars

* add hadoop 3.2.1 checksum and remove azure_vars

- integrate default value handling for host_var and play_var in to base
  config class accessor for respective var dictionaries
- remove use of 'azure_vars' and consolidate as host_vars

* remove fstype and force_format from base, existing, and azure config

* remove checks for fstype and force_format from existing and azure tests

* addressing review comments

- made config.shutdown_delay_minutes not abstract so existing and azure
  cluster types wouldn't have to implement it
- changed _ansible_vars dictionary values a list of structs rather than
  tuples so that each field has an understandable name
- removed instance_tags from base config
